move components and styling files to Components files
