+++ src/Components/Error/Error.js
@@ -1,8 +1,7 @@
 import React from "react";
-import { useLocation } from 'react-router-dom'
 import PropTypes from 'prop-types';
 
-import './Error.css'
+import './Error.css';
 
 const Error = ({ dismissError }) => {
