+++ src/Components/Options/Options.css
@@ -55,7 +55,6 @@
 
   .play-image,
   .fav-image {
-    /* height: 50%; */
     width: 50%;
     margin: 20px;
   }
