+++ src/Components/PageNotFound/PageNotFound.js
@@ -1,6 +1,7 @@
 import React from "react";
-import { Link } from "react-router-dom"
-import './PageNotFound.css'
+import { Link } from "react-router-dom";
+
+import './PageNotFound.css';
 
 const PageNotFound = () => {
